Ver_0/requirements-not-cloud-safe.txt
# ============================
# CruncherX — Unified Requirements (Stable)
# Cloud Engine + Local Bulldozer Engine
# ============================

# --- Core App ---
streamlit==1.32.0
python-dotenv==1.0.1

# --- Supabase (Stable versions for Python 3.11) ---
supabase==2.3.5
gotrue==2.4.2
postgrest==0.15.0
realtime==1.0.0
storage3==0.5.3
httpx==0.24.1

# --- PDF Handling (Cloud-Safe) ---
pypdf==4.2.0
pdf2image==1.17.0
Pillow==10.3.0
PyMuPDF==1.24.9
pytesseract==0.3.10

# --- Compression Helpers ---
numpy==1.26.4

# --- Optional Local Bulldozer Engine ---
pikepdf==8.13.0
reportlab==4.1.0

# --- Utilities ---
requests==2.31.0
urllib3==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6

# --- Streamlit Extensions ---
st-social-media-links==0.1.5

# --- JSON & Schema ---
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
referencing==0.34.0
rpds-py==0.18.0
MarkupSafe==2.1.5
Jinja2==3.1.3

# --- Git Integration ---
GitPython==3.1.42
gitdb==4.0.11
smmap==5.0.1

# --- Time & Date ---
python-dateutil==2.9.0.post0
pytz==2024.1
tzdata==2024.1

# --- Visualization ---
altair==5.2.0
plotly==5.18.0
pydeck==0.9.1

# --- Packaging ---
packaging==23.2
six==1.16.0
